Moves visual button configuration to the view from the directive.
